+++ bigchaindb/common/utils.py
