+++ bigchaindb/web/server.py
@@ -9,7 +9,7 @@ import multiprocessing
 from flask import Flask
 import gunicorn.app.base
 
-from bigchaindb import util
+from bigchaindb import utils
 from bigchaindb import Bigchain
 from bigchaindb.web.views.info import info_views
 from bigchaindb.web.views.transactions import transaction_views
@@ -66,7 +66,7 @@ def create_app(*, debug=False, threads=4):
 
     app.debug = debug
 
-    app.config['bigchain_pool'] = util.pool(Bigchain, size=threads)
+    app.config['bigchain_pool'] = utils.pool(Bigchain, size=threads)
     app.config['monitor'] = Monitor()
 
     app.register_blueprint(info_views, url_prefix='/')
