+++ tests/test_consensus.py
@@ -2,7 +2,7 @@
 def test_verify_vote_passes(b, structurally_valid_vote):
     from bigchaindb.consensus import BaseConsensusRules
     from bigchaindb.common import crypto
-    from bigchaindb.common.util import serialize
+    from bigchaindb.common.utils import serialize
     vote_body = structurally_valid_vote['vote']
     vote_data = serialize(vote_body)
     signature = crypto.PrivateKey(b.me_private).sign(vote_data.encode())
@@ -28,7 +28,7 @@ def test_verify_vote_fails_signature(b, structurally_valid_vote):
 def test_verify_vote_fails_schema(b):
     from bigchaindb.consensus import BaseConsensusRules
     from bigchaindb.common import crypto
-    from bigchaindb.common.util import serialize
+    from bigchaindb.common.utils import serialize
     vote_body = {}
     vote_data = serialize(vote_body)
     signature = crypto.PrivateKey(b.me_private).sign(vote_data.encode())
